server/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
gunicorn==20.0.4
requests==2.25.1
djangorestframework==3.12.2
pytesseract==0.3.7
pillow==8.2.0
django-filter==2.4.0
